+++ CHANGELOG.md
@@ -9,6 +9,12 @@ his file's format is define in
 and the release workflow reads it to set github's release notes.
 
 
+## Unreleased
+
+### Added
+
+- Send no more than one email a minute to a user
+-
 ## [0.3.1] 2020-5-22
 
 ### Fixed